+++ LapsTeamTwo/src/main/java/com/teamtwo/laps/controller/AdminUserController.java
@@ -3,6 +3,7 @@ package com.teamtwo.laps.controller;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,21 +22,13 @@ import com.teamtwo.laps.model.User;
 import com.teamtwo.laps.service.StaffMemberService;
 import com.teamtwo.laps.service.UserService;
 
-
-
-
-
-
-
-@RequestMapping(value="/admin/user")
+@RequestMapping(value = "/admin/user")
 @Controller
 public class AdminUserController {
 
-
-
 	@Autowired
 	private UserService uService;
-	
+
 	@Autowired
 	private StaffMemberService smService;
 
@@ -45,23 +38,50 @@ public class AdminUserController {
 	 * @return
 	 */
 
-
 	@RequestMapping(value = "/list", method = RequestMethod.GET)
-	public ModelAndView userListPage() {
-		ModelAndView mav = new ModelAndView("user-list");
-		ArrayList<User> userList = uService.findAllUsers();
-		mav.addObject("userList", userList);
+	public ModelAndView userListPage(HttpSession session) {
+
+		ModelAndView mav = new ModelAndView("login");
+		try {
+			UserSession us = (UserSession) session.getAttribute("USERSESSION");
+			if (us.getSessionId() != null && us.getUser().getIsAdmin()) {
+
+				mav = new ModelAndView("user-list");
+				ArrayList<User> userList = uService.findAllUsers();
+				mav.addObject("userList", userList);
+			} else {
+				mav = new ModelAndView("unauthorized-admin-access");
+			}
+
+		} catch (NullPointerException e) {
+			// TODO: handle exception
+			mav = new ModelAndView("unauthorized-access");
+		}
 		return mav;
 	}
-	
+
 	@RequestMapping(value = "/create", method = RequestMethod.GET)
-	public ModelAndView newUserPage() {
-		ModelAndView mav = new ModelAndView("user-new", "user", new User());
-		ArrayList<Integer> sidList = smService.findAllStaffId();
-		mav.addObject("sidlist", sidList);
+	public ModelAndView newUserPage(HttpSession session) {
+
+		ModelAndView mav = new ModelAndView("login");
+		try {
+			UserSession us = (UserSession) session.getAttribute("USERSESSION");
+			if (us.getSessionId() != null && us.getUser().getIsAdmin()) {
+
+				mav = new ModelAndView("user-new", "user", new User());
+				ArrayList<Integer> sidList = smService.findAllStaffId();
+				mav.addObject("sidlist", sidList);
+			} else {
+				mav = new ModelAndView("unauthorized-admin-access");
+			}
+
+		} catch (NullPointerException e) {
+			// TODO: handle exception
+			mav = new ModelAndView("unauthorized-access");
+		}
 		return mav;
 	}
-	
+
 	@RequestMapping(value = "/create", method = RequestMethod.POST)
 	public ModelAndView createNewUser(@ModelAttribute @Valid User user, BindingResult result,
 			final RedirectAttributes redirectAttributes) {
@@ -78,20 +98,34 @@ public class AdminUserController {
 		redirectAttributes.addFlashAttribute("message", message);
 		return mav;
 	}
-	
+
 	@RequestMapping(value = "/edit/{id}", method = RequestMethod.GET)
-	public ModelAndView editUserPage(@PathVariable String id) {
-		ModelAndView mav = new ModelAndView("user-edit");
-		User user = uService.findUser(id);
-		mav.addObject("user", user);
-		ArrayList<Integer> sidList = smService.findAllStaffId();
-		mav.addObject("sidlist", sidList);
+	public ModelAndView editUserPage(@PathVariable String id, HttpSession session) {
+
+		ModelAndView mav = new ModelAndView("login");
+		try {
+			UserSession us = (UserSession) session.getAttribute("USERSESSION");
+			if (us.getSessionId() != null && us.getUser().getIsAdmin()) {
+
+				mav = new ModelAndView("user-edit");
+				User user = uService.findUser(id);
+				mav.addObject("user", user);
+				ArrayList<Integer> sidList = smService.findAllStaffId();
+				mav.addObject("sidlist", sidList);
+			} else {
+				mav = new ModelAndView("unauthorized-admin-access");
+			}
+
+		} catch (NullPointerException e) {
+			// TODO: handle exception
+			mav = new ModelAndView("unauthorized-access");
+		}
 		return mav;
 	}
-	
+
 	@RequestMapping(value = "/edit/{id}", method = RequestMethod.POST)
 	public ModelAndView editUser(@ModelAttribute @Valid User user, BindingResult result, @PathVariable String id,
-			final RedirectAttributes redirectAttributes){
+			final RedirectAttributes redirectAttributes) {
 
 		if (result.hasErrors())
 			return new ModelAndView("user-edit");
@@ -106,15 +140,28 @@ public class AdminUserController {
 	}
 
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
-	public ModelAndView deleteUser(@PathVariable String id, final RedirectAttributes redirectAttributes)
-	{
+	public ModelAndView deleteUser(@PathVariable String id, final RedirectAttributes redirectAttributes,
+			HttpSession session) {
 
-		ModelAndView mav = new ModelAndView("redirect:/admin/user/list");
-		User user = uService.findUser(id);
-		uService.removeUser(user);
-		String message = "The user " + user.getUserId() + " was successfully deleted.";
+		ModelAndView mav = new ModelAndView("login");
+		try {
+			UserSession us = (UserSession) session.getAttribute("USERSESSION");
+			if (us.getSessionId() != null && us.getUser().getIsAdmin()) {
 
-		redirectAttributes.addFlashAttribute("message", message);
+				mav = new ModelAndView("redirect:/admin/user/list");
+				User user = uService.findUser(id);
+				uService.removeUser(user);
+				String message = "The user " + user.getUserId() + " was successfully deleted.";
+
+				redirectAttributes.addFlashAttribute("message", message);
+			} else {
+				mav = new ModelAndView("unauthorized-admin-access");
+			}
+
+		} catch (NullPointerException e) {
+			// TODO: handle exception
+			mav = new ModelAndView("unauthorized-access");
+		}
 		return mav;
 	}
